requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.2
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloud-sql-python-connector==1.2.1
cloudevents==1.9.0
cryptography==40.0.1
deprecation==2.1.0
Flask==2.2.3
flask-oauth2-validation=0.1.2
flask-oauth2-validation=0.1.2
frozenlist==1.3.3
functions-framework==3.3.0
google-auth==2.16.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
packaging==23.0
pg8000==1.29.4
pluggy==1.0.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyMySQL==1.0.2
pytest==7.2.2
python-dateutil==2.8.2
requests==2.28.2
rsa==4.9
scramp==1.4.4
six==1.16.0
SQLAlchemy==2.0.7
typing_extensions==4.5.0
urllib3==1.26.15
watchdog==3.0.0
Werkzeug==2.2.3
yarl==1.8.2
